+++ core/src/main/java/com/adobe/aem/commons/assetshare/content/properties/impl/TagTitlesImpl.java
@@ -35,6 +35,7 @@ import org.osgi.service.metatype.annotations.Designate;
 import org.osgi.service.metatype.annotations.ObjectClassDefinition;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 
@@ -81,6 +82,8 @@ public class TagTitlesImpl extends AbstractComputedProperty<List<String>> {
                 tagLabels.add(tag.getTitle(locale));
             }
         }
+
+        Collections.sort(tagLabels);
 
         return tagLabels;
     }